CXpcModel: fix: error value (unsigned -1) was treated as a valid versionNumber
git-svn-id: http://b035a0a.online-server.cloud/svn/AcfPublic@515 7b0a5285-9844-ea4a-88f9-6912f79b293c

+++ Acf/Include/icomp/CXpcModel.cpp
@@ -141,8 +141,8 @@ bool CXpcModel::Serialize(iser::IArchive& archive)
 	quint32 versionNumber = 0;
 	const iser::IVersionInfo& versionInfo = archive.GetVersionInfo();
 	versionInfo.GetVersionNumber(0, versionNumber);
-
-	if (versionNumber > 2473){
+	
+	if ((signed int)versionNumber > 2473){  // cast to signed because of error value (-1)
 		//Registry files available
 
 		int registryFilesCount = 0;
